refactor: fix bugs
