+++ src/main/java/ru/yandex/practicum/filmorate/repository/film/FilmRepositoryImpl.java
@@ -1,6 +1,6 @@
 package ru.yandex.practicum.filmorate.repository.film;
 
-import ru.yandex.practicum.filmorate.repository.genres.GenresRepositoryImpl;
+import org.springframework.jdbc.core.BatchPreparedStatementSetter;
 import ru.yandex.practicum.filmorate.entity.Genre;
 import org.springframework.jdbc.core.JdbcTemplate;
 import ru.yandex.practicum.filmorate.entity.Film;
@@ -8,12 +8,12 @@ import ru.yandex.practicum.filmorate.entity.Mpa;
 import org.springframework.stereotype.Component;
 import lombok.RequiredArgsConstructor;
 
+import java.util.function.Function;
 import java.util.stream.Collectors;
+import java.sql.PreparedStatement;
 import java.sql.SQLException;
 import java.sql.ResultSet;
-import java.util.Optional;
-import java.util.Objects;
-import java.util.List;
+import java.util.*;
 
 /**
  * @author Oleg Khilko
@@ -22,7 +22,6 @@ import java.util.List;
 @Component
 @RequiredArgsConstructor
 public class FilmRepositoryImpl implements FilmRepository {
-    private final GenresRepositoryImpl repository;
     private final JdbcTemplate jdbcTemplate;
 
     @Override
@@ -42,7 +41,7 @@ public class FilmRepositoryImpl implements FilmRepository {
                     .map(Genre::getId)
                     .collect(Collectors.toList());
 
-            repository.addGenres(film.getId(), genreIds);
+            addGenres(film.getId(), genreIds);
         }
         return film;
     }
@@ -60,19 +59,19 @@ public class FilmRepositoryImpl implements FilmRepository {
                 film.getMpa().getId(),
                 film.getId());
 
-        var filmGenres = repository.getFilmGenres(film.getId());
+        var filmGenres = getFilmGenres(film.getId());
 
         var filmsAdded = film.getGenres().stream()
                 .filter(genre -> !filmGenres.contains(genre))
                 .map(Genre::getId)
                 .collect(Collectors.toList());
-        repository.addGenres(film.getId(), filmsAdded);
+        addGenres(film.getId(), filmsAdded);
 
         var filmsRemoved = filmGenres.stream()
                 .filter(genre -> !film.getGenres().contains(genre))
                 .map(Genre::getId)
                 .collect(Collectors.toList());
-        repository.removeGenres(film.getId(), filmsRemoved);
+        removeGenres(film.getId(), filmsRemoved);
 
         return film;
     }
@@ -97,13 +96,13 @@ public class FilmRepositoryImpl implements FilmRepository {
             var film = Film.builder()
                     .releaseDate(Objects.requireNonNull(filmRow.getDate("release_date")).toLocalDate())
                     .mpa(new Mpa(filmRow.getInt("mpa"), filmRow.getString("MPA_NAME")))
+                    .description(Objects.requireNonNull(filmRow.getString("description")))
                     .name(Objects.requireNonNull(filmRow.getString("name")))
-                    .description(filmRow.getString("description"))
                     .duration(filmRow.getInt("duration"))
                     .id(filmRow.getInt("id"))
                     .build();
 
-            film.getGenres().addAll(repository.getFilmGenres(filmId));
+            film.getGenres().addAll(getFilmGenres(filmId));
             film.getLikes().addAll(getUserLikes(filmId));
             return Optional.of(film);
         } else {
@@ -136,10 +135,73 @@ public class FilmRepositoryImpl implements FilmRepository {
                         "LIMIT ?", (resultSet, rowNum) -> mapFilm(resultSet), count);
 
         films.forEach(film -> {
-            film.getGenres().addAll(repository.getFilmGenres(film.getId()));
+            film.getGenres().addAll(getFilmGenres(film.getId()));
             film.getLikes().addAll(getUserLikes(film.getId()));
         });
         return films;
+    }
+
+    @Override
+    public Map<Integer, Genre> getGenresByIds(List<Integer> ids) {
+        return jdbcTemplate.query(String.format(
+                                "SELECT * FROM genre WHERE id IN (%s)",
+                                String.join(",", Collections.nCopies(ids.size(), "?"))),
+                        (resultSet, rowNum) -> mapGenre(resultSet), ids.toArray())
+                .stream()
+                .collect(Collectors.toMap(Genre::getId, Function.identity()));
+    }
+
+    private List<Integer> getUserLikes(int filmId) {
+        return jdbcTemplate.query("SELECT user_id " +
+                "FROM film_likes " +
+                "WHERE film_id = ?", (resultSet, rowNum) ->
+                resultSet.getInt("user_id"), filmId);
+    }
+
+    private void addGenres(int filmId, List<Integer> genres) {
+        jdbcTemplate.batchUpdate(
+                "INSERT INTO film_genre (film_id, genre_id) VALUES (?,?)",
+                new BatchPreparedStatementSetter() {
+                    public void setValues(PreparedStatement preparedStatement, int value)
+                            throws SQLException {
+                        preparedStatement.setInt(1, filmId);
+                        preparedStatement.setInt(2, genres.get(value));
+                    }
+
+                    public int getBatchSize() {
+                        return genres.size();
+                    }
+                });
+    }
+
+    private void removeGenres(int filmId, List<Integer> genres) {
+        jdbcTemplate.batchUpdate(
+                "DELETE FROM film_genre WHERE film_id = ? AND genre_id = ?",
+                new BatchPreparedStatementSetter() {
+                    public void setValues(PreparedStatement ps, int value)
+                            throws SQLException {
+                        ps.setInt(1, filmId);
+                        ps.setInt(2, genres.get(value));
+                    }
+
+                    public int getBatchSize() {
+                        return genres.size();
+                    }
+                });
+    }
+
+    private List<Genre> getFilmGenres(int filmId) {
+        return jdbcTemplate.query(
+                "SELECT * " +
+                        "FROM genre " +
+                        "INNER JOIN film_genre ON film_genre.genre_id = genre.id " +
+                        "AND film_genre.film_id = ?",
+                (resultSet, rowNum) -> mapGenre(resultSet), filmId);
+    }
+
+    private Genre mapGenre(ResultSet resultSet) throws SQLException {
+        return new Genre(resultSet.getInt("id"),
+                resultSet.getString("name"));
     }
 
     private Film mapFilm(ResultSet resultSet) throws SQLException {
@@ -154,16 +216,9 @@ public class FilmRepositoryImpl implements FilmRepository {
                 .name(resultSet.getString("name"))
                 .build();
 
-        film.getGenres().addAll(repository.getFilmGenres(id));
+        film.getGenres().addAll(getFilmGenres(id));
         film.getLikes().addAll(getUserLikes(id));
 
         return film;
-    }
-
-    private List<Integer> getUserLikes(int filmId) {
-        return jdbcTemplate.query("SELECT user_id " +
-                "FROM film_likes " +
-                "WHERE film_id = ?", (resultSet, rowNum) ->
-                resultSet.getInt("user_id"), filmId);
     }
 }
