+++ src/main/java/ru/yandex/practicum/filmorate/service/user/UserServiceImpl.java
@@ -1,8 +1,8 @@
 package ru.yandex.practicum.filmorate.service.user;
 
 import ru.yandex.practicum.filmorate.repository.user.UserRepository;
-import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
+import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.entity.User;
 import org.springframework.stereotype.Service;
 import lombok.RequiredArgsConstructor;
